Fix: Allow the Slider Menu and the Hamburger Menu to work in GitHub Pages
-Using href="/" breaks the slider, since it does not show the animation
properly and also creates the Page 404 not found error while using
href="#" makes it not possible to exit the hamburger menu. Created
toggleAnchorMenuEvent function to solve this problem.

NOTE: toggleAnchorMenuEvent() is a hack to make my hamburger menu and slider
menu work together in GitHub pages.

1 - To properly work, my slider menu requires the page to not reload (href="#")
The anchors in my hamburger menu reload the page to exit it (href="/")
2 - In GitHub Pages anchors with href="/" don't work as it redirects you
to [username].github.io instead of [user_name].github.io/[repo_name].

If you are not using GitHub Pages AND not using the slider menu in conjunction
with the hamburger menu remove toggleAnchorMenuEvent().